backend/requirements.txt
fastapi
uvicorn
spotipy
python-dotenv
httpx
